fix(frontend): the agent header keeps the app's commit adapter and waits for a verdict
Moving the agent config bar into the package took its commit button with it, and the
OSS host stopped passing `appId`/`onAfterCommit`/`onCommitted` — so an agent commit
skipped the registry and evaluator cache refresh and the onboarding event, leaving
both lists stale. The adapter those props carry is now a hook, so the button wrapper
and the agent header share one definition instead of the header re-deriving it.

The page header read `isAgentWorkflow`, which is false while agent-ness is merely
UNKNOWN, so the title flashed "Playground" on every agent reload before swapping to
the agent identity. It waits for the same confirmed signal the eval action stack
already waits for.
